Version: 2001-06-01 ksh93l
Sourced from ast-open-history version 2001-06-01 and:
1f439a39a660f7fa5c4dcfe1441971f7532cdc95f475583d663d99e826674ece  INIT.2001-01-01b.tgz

+++ src/lib/libast/sfio/sfdlen.c
@@ -52,7 +52,7 @@ Sfdouble_t	v;
 
 	for(w = 1; w <= N_ARRAY; ++w)
 	{	/* get 2^SF_PRECIS precision at a time */
-		n = (int)(x = ldexp(v,SF_PRECIS));
+		n = (int)(x = ldexpl(v,SF_PRECIS));
 		v = x-n;
 		if(v <= 0.)
 			break;
